+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Sniffers can now dig out golden dandelions
 - More utility for the clock
+- Golden dandelion particles
+- Baby mobs follow players holding golden dandelion when their age is not locked
 
 ## [3.2.1] - 2026-03-26
 ### Fixed